requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
assert-methods==0.1.8
attrs==21.2.0
beeprint==2.4.10
Faker==6.6.3
graphql-core==3.1.6
graphqlapiobject==3.1.2
iniconfig==1.1.1
jmespath==0.10.0
packaging==21.2
pluggy==1.0.0
py==1.10.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-utils==2.5.6
sgqlc==13.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
urwid==2.1.2
